# src/main/java/com/diegchav/staywise/api/dto/CreateHotelRequest.java
package com.diegchav.staywise.api.dto;

import jakarta.validation.constraints.NotBlank;

public record CreateHotelRequest(
        @NotBlank(message = "Name is required")
        String name,

        @NotBlank(message = "Address is required")
        String address,

        @NotBlank(message = "City is required")
        String city,

        @NotBlank(message = "Country is required")
        String country
) {}