# src/main/resources/static/openapi.yaml
openapi: 3.1.1
info:
  title: Booking Service API
  version: 0.1.0
  description: Hotel booking service API

servers:
  - url: http://localhost:8080

paths:
  /api/availability:
    get:
      summary: Search room availability
      description: Returns available room types for a hotel within a date range.
      operationId: searchAvailability
      tags: [Search]

      parameters:
        - name: hotelId
          in: query
          required: true
          schema:
            type: string
            format: uuid

        - name: checkIn
          in: query
          required: true
          schema:
            type: string
            format: date

        - name: checkOut
          in: query
          required: true
          schema:
            type: string
            format: date

      responses:
        "200":
          description: Available room types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailableRoomType"
        "400":
          description: Invalid date range

  /api/bookings:
    post:
      summary: Create a booking
      operationId: createBooking
      tags: [Bookings]

      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
          description: Unique key to make booking request idempotent

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'

      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        "400":
          description: Invalid request
        "409":
          description: No availability

    get:
      summary: List bookings by user
      operationId: listBookings
      tags: [Bookings]

      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingResponse'

  /api/bookings/{bookingId}:
    get:
      summary: Get booking by ID
      operationId: getBooking
      tags: [Bookings]

      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        "404":
          description: Booking not found

  /api/bookings/{bookingId}/cancel:
    post:
      summary: Cancel a booking
      description: |
        Cancels an existing booking and releases reserved inventory.
        This operation is idempotent — cancelling an already cancelled booking
        returns success without additional side effects.
      operationId: cancelBooking
      tags: [Bookings]

      parameters:
        - name: bookingId
          in: path
          required: true
          description: Unique identifier of the booking
          schema:
            type: string
            format: uuid

      responses:
        "204":
          description: Booking successfully cancelled

        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "409":
          description: Booking cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/hotels:
    post:
      summary: Create a hotel
      operationId: createHotel
      tags: [Hotels]

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHotelRequest"

      responses:
        "201":
          description: Hotel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelResponse"

        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/ValidationError"

    get:
      summary: List all hotels
      operationId: listHotels
      tags: [Hotels]

      responses:
        "200":
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/HotelResponse"

  /api/hotels/{hotelId}:
    get:
      summary: Get hotel by ID
      operationId: getHotel
      tags: [Hotels]

      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "200":
          description: Hotel found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelResponse"

        "404":
          description: Not found

    patch:
      summary: Partially update a hotel
      operationId: updateHotel
      tags: [Hotels]

      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHotelRequest"

      responses:
        "200":
          description: Hotel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelResponse"

        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/ValidationError"

        "404":
          description: Not found

    delete:
      summary: Delete a hotel
      operationId: deleteHotel
      tags: [Hotels]

      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "204":
          description: Hotel deleted

components:
  schemas:
    AvailableRoomType:
      type: object
      properties:
        roomTypeId:
          type: string
          format: uuid
        availableRooms:
          type: integer

    CreateBookingRequest:
      type: object
      required:
        - userId
        - hotelId
        - roomTypeId
        - checkIn
        - checkOut
      properties:
        userId:
          type: string
          format: uuid
        hotelId:
          type: string
          format: uuid
        roomTypeId:
          type: string
          format: uuid
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date

    BookingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        hotelId:
          type: string
          format: uuid
        roomTypeId:
          type: string
          format: uuid
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        status:
          type: string
          example: CONFIRMED
        totalPrice:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time

    CreateHotelRequest:
      type: object
      required:
        - name
        - address
        - city
        - country
      properties:
        name:
          type: string
          example: "Hilton Reforma"
        address:
          type: string
          example: "Reforma 101"
        city:
          type: string
          example: "Mexico City"
        country:
          type: string
          example: "Mexico"

    UpdateHotelRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string

    HotelResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        city:
          type: string
        country:
          type: string
        rating:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: integer

    ValidationError:
      type: object
      properties:
        field:
          type: string
        error:
          type: string